requirements.txt
altgraph==0.17.2
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.5.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachelib==0.1.1
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.2.14
django-cors-headers==3.13.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
importlib-metadata==4.12.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.9.0
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.21.6
packaging==20.9
pandas==1.3.5
pandocfilters==1.4.3
parso==0.8.1
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.1.1
prometheus-client==0.9.0
prompt-toolkit==3.0.16
pycparser==2.20
pygame==2.0.1
Pygments==2.8.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.2.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.15
sqlparse==0.4.2
swapper==1.3.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==4.3.0
virtualenv==20.4.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zipp==3.8.1
zope.interface==5.4.0
